Read 작업 완료 관련 백 커밋 VO > BNum 타입 변경 (Int > Long)에 따른 getBoard 파라미터 타입 변경 (+ Service) Repository > getBoard 기능 개발 Controller > Insert 기능 추가, View 타입 변경 및 테스트 로그 추가

+++ src/main/java/com/springboot/react/cboard/CBoardService.java
@@ -60,7 +60,7 @@ public class CBoardService {
 	   return cboardDAO.getPagingBoard(pageNum);
    }
    
-   public ResponseEntity<Map> getBoard(Integer bnum){
+   public ResponseEntity<Map> getBoard(Long bnum){
 	   return cboardDAO.getBoard(bnum);
    }
